2019 FIRST Destination: Deep Space (#56) Initial Commit
Master is now development branch for v2019, which is not compatible with any other year's game. 
Development on v2018 is now finished and no further support will be provided. Still much to be done with CouchDB automation and developing StrangeBox hardware, but we now have a working prerelease codebase that is usable for DDS. Needless to say, this is still prerelease code, use at your own risk.